refact: refatoramento da arquitetura do endpoint de posts para usar controller e business
